feat: implement cart context and provider for managing cart items with AsyncStorage persistence

+++ src/contexts/cartContext.tsx
@@ -1,7 +1,7 @@
 import { createContext, ReactNode, useContext, useEffect, useState } from 'react';
 import { Product } from '../types/product';
 import AsyncStorage from '@react-native-async-storage/async-storage';
-import { ToastAndroid } from 'react-native';
+import { Alert, ToastAndroid } from 'react-native';
 
 interface CartContextType {
     cartItems: Product[];
@@ -54,31 +54,55 @@ export const CartProvider = ({ children }: { children: ReactNode }) => {
     };
 
     const removeFromCart = (id: number) => {
-        try {
-            setLoading(true);
-            AsyncStorage.setItem(PRODUCT_KEY, JSON.stringify(cartItems.filter((item) => item.id !== id)));
-            setCartItems((prevItems) => prevItems.filter((item) => item.id !== id));
-        } catch (error) {
-            console.error('Failed to remove product from cart:', error);
-            setError('Failed to remove product from cart');
-        } finally {
-            setLoading(false);
-            ToastAndroid.show('Product removed from cart', ToastAndroid.SHORT);
-        }
+        Alert.alert('Remove product', 'Are you sure want to remove product?', [
+            {
+                text: 'Cancel',
+                style: 'cancel'
+            },
+            {
+                text: 'Remove',
+                style: 'destructive',
+                onPress: () => {
+                    try {
+                        setLoading(true);
+                        AsyncStorage.setItem(PRODUCT_KEY, JSON.stringify(cartItems.filter((item) => item.id !== id)));
+                        setCartItems((prevItems) => prevItems.filter((item) => item.id !== id));
+                    } catch (error) {
+                        console.error('Failed to remove product from cart:', error);
+                        setError('Failed to remove product from cart');
+                    } finally {
+                        setLoading(false);
+                        ToastAndroid.show('Product removed from cart', ToastAndroid.SHORT);
+                    }
+                }
+            }
+        ])
     };
 
     const clearCart = () => {
-        try {
-            setLoading(true);
-            AsyncStorage.removeItem(PRODUCT_KEY);
-            setCartItems([]);
-        } catch (error) {
-            console.error('Failed to clear cart:', error);
-            setError('Failed to clear cart');
-        } finally {
-            setLoading(false);
-            ToastAndroid.show('Cart cleared', ToastAndroid.SHORT);
-        }
+        Alert.alert('Clear cart', 'Are you sure you want to clear cart?', [
+            {
+                text: 'Cancel',
+                style: 'cancel'
+            },
+            {
+                text: 'Clear',
+                style: 'destructive',
+                onPress: () => {
+                    try {
+                        setLoading(true);
+                        AsyncStorage.removeItem(PRODUCT_KEY);
+                        setCartItems([]);
+                    } catch (error) {
+                        console.error('Failed to clear cart:', error);
+                        setError('Failed to clear cart');
+                    } finally {
+                        setLoading(false);
+                        ToastAndroid.show('Cart cleared', ToastAndroid.SHORT);
+                    }
+                }
+            }
+        ])
     };
 
     return (
